+++ plugins/gui/include/gui/context_manager_widget/context_manager_widget.h
@@ -134,7 +134,6 @@ namespace hal
         QString mOpenIconPath;
         QString mOpenIconStyle;
 
-        QAction* mSearchAction;
         QString mSearchIconPath;
         QString mSearchIconStyle;