+++ plugins/gui/include/gui/selection_details_widget/gate_details_widget.h
@@ -74,6 +74,8 @@ namespace hal
          */
         void update(const u32 gate_id);
 
+        void hideSectionsWhenEmpty(bool hide) override;
+
     public Q_SLOTS:
 
         /**