+++ plugins/gui/include/gui/selection_details_widget/net_details_widget.h
@@ -78,6 +78,8 @@ namespace hal
          */
         void update(u32 mNetId);
 
+        void hideSectionsWhenEmpty(bool hide) override;
+
     public Q_SLOTS:
 
         /**