+++ plugins/gui/include/gui/settings/settings_widgets/settings_widget.h
@@ -23,8 +23,6 @@
 
 #pragma once
 
-#include "gui/settings/preview_widget.h"
-
 #include <QVariant>
 #include <QToolButton>
 #include <QBoxLayout>
@@ -38,72 +36,62 @@ class QVBoxLayout;
 
 namespace hal
 {
+    class SettingsItem;
+
     class SettingsWidget : public QFrame
     {
         Q_OBJECT
         Q_PROPERTY(QColor highlightColor READ highlightColor WRITE setHighlightColor)
         Q_PROPERTY(bool dirty READ dirty WRITE setDirty)
-        Q_PROPERTY(bool conflicts READ conflicts WRITE setConflicts)
 
     public:
-        enum class preview_position
-        {
-            bottom = 0,
-            right = 1
-        };
-
-        explicit SettingsWidget(const QString& key, QWidget* parent = 0);
+        explicit SettingsWidget(SettingsItem* item, QWidget* parent = 0);
 
         QColor highlightColor();
-        QString key();
         void setHighlightColor(const QColor& color);
 
-        void resetLabels();
-        bool matchLabels(const QString& string);
+        void reset_labels();
+        bool matchLabel(const QString& needle);
 
         bool dirty() const;
-        void prepare(const QVariant& value, const QVariant& default_value);
-        void markSaved();
-        void setConflicts(bool conflicts);
-        bool conflicts() const;
+        void prepare();
 
-        void setPreviewWidget(PreviewWidget* widget);
-        void setPreviewPosition(preview_position position);
+        const SettingsItem* settingsItem() const { return mSettingsItem; }
+        virtual void acceptValue();
+        virtual void loadCurrentValue();
 
         virtual void load(const QVariant& value) = 0;
-        virtual QVariant value()                  = 0;
+        virtual QVariant value()                 = 0;
+        virtual void clearEditor() {;}
 
     public Q_SLOTS:
-        void handleRollback();
-        void handleReset();
+        virtual void handleRevertModification();
+        virtual void handleSetDefaultValue(bool setAll=false);
+        void handleItemDestroyed();
 
     Q_SIGNALS:
-        void settingUpdated(SettingsWidget* sender, const QString& key, const QVariant& value);
+        void valueChanged();
+        void triggerDescriptionUpdate(SettingsItem* item);
+        void triggerRemoveWidget(SettingsWidget* widget);
 
     protected:
         void setDirty(bool dirty);
-        void triggerSettingUpdated();
+        void trigger_setting_updated();
         void enterEvent(QEvent* event);
         void leaveEvent(QEvent* event);
 
-        QVBoxLayout* mLayout;
+        QVBoxLayout* m_layout;
         QBoxLayout* mContainer;
-        QHBoxLayout* mTopBar;
-        QLabel* mName;
-        QToolButton* mRevert;
-        QToolButton* mDefault;
+        QHBoxLayout* m_top_bar;
+        QLabel* mNameLabel;
+        QToolButton* mRevertButton;
+        QToolButton* mDefaultButton;
 
-        QList<QPair<QLabel*, QString>> mLabels;
+        QList<QPair<QLabel*, QString>> m_labels;
+        SettingsItem* mSettingsItem;
 
     private:
         QColor mHighlightColor;
-        QString mKey;
-        bool mSignalsEnabled = true;
-        bool mPrepared = false;
         bool mDirty = false;
-        bool mConflicts = false;
-        QVariant mLoadedValue;
-        QVariant mDefaultValue;
-        PreviewWidget* m_preview = nullptr;
     };
 }