+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* theme setting (and many other settings) can be changed during hal session
+* warning message when attempting to assingn a keybinding already in use
+* incremental search for settings enabled
+* enable/disable extra window to list all executed instances from UserAction
 * fixed drag'n drop bug related to negative coordinates
 * added user action system to enable recording and reverting actions within the GUI
   * moved most GUI actions to the new user action system, including interactions with the graph view and view management